+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 - Add `transform` option to attributes
 - Add option to output failure backtraces
+- Add option to handle exceptions for non bang exceptions
 - Add durability (retries) to execution
 
 ### Changes